+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.8.0] / 2024-12-16
 ### Features
 - Make project public
+### Application
+- Remove `net46` framework, remove `Revit 2017` and `Revit 2018` support.
 ### Build
 - Update `ci` to use public `nuget` packages.