+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove `net46` framework, remove `Revit 2017` and `Revit 2018` support.
 ### Command
 - Fix `AppUtils` to show `ProductName`.
+- Update `ProcessStart` to show `arguments` length in debug.
 ### Console
 - Update `ricaun.Revit.Installation` to `1.3.1`.
 - Add `EnvironmentVariable` class, with `ProcessArguments` and `TimeoutNotBusyMaxSeconds` variables.